Prepare release of version 1.6.1
- Based on SQLite version 3.41.1

Symbol MAX_PATHNAME (used on Unix-like platforms) has a fixed value of 512 in the original SQLite source code. This can now be configured at compile time to use a higher value (like 4096 - which is supported by most Linux variants) (see issue #104). Use symbol SQLITE3MC_MAX_PATHNAME to define a higher value.

+++ src/sqlite3mc_vfs.c
@@ -1247,7 +1247,7 @@ sqlite3mcCheckVfs(const char* zVfs)
   if (pVfs == NULL)
   {
     /* VFS not found */
-    int prefixLen = strlen(SQLITE3MC_VFS_NAME);
+    int prefixLen = (int) strlen(SQLITE3MC_VFS_NAME);
     if (strncmp(zVfs, SQLITE3MC_VFS_NAME, prefixLen) == 0)
     {
       /* VFS name starts with prefix. */
